Fix VPAID clickThrough in iOS, android chrome broken now
