+++ src/helpers/loadVPAID.js
@@ -8,11 +8,12 @@ export default function loadVPAID(url, container) {
             'style',
             'border:0px;margin:0px;opacity:1;padding:0px;height:100%;position:absolute;width:100%;top:0;left:0;'
         );
-        container.prepend(iframe);
+        container.appendChild(iframe);
         // url points to the ad js file
         iframe.contentWindow.document.write(
             // split the end script tag to prevent closing js prematurely
-            `<head></head><body><script id="${SCRIPT_ID}" src="${url}" async></scr` + 'ipt></body>'
+            `<head><style>body{margin:0}</style></head><body><script id="${SCRIPT_ID}" src="${url}" async></scr` +
+                'ipt></body>'
         );
 
         const script = iframe.contentWindow.document.getElementById(SCRIPT_ID);
@@ -24,13 +25,13 @@ export default function loadVPAID(url, container) {
         }
 
         // or load script
-        script.onload = () => {
+        script.addEventListener('load', () => {
             VPAIDCreative = getVPAIDCreative(iframe);
             resolve({ VPAIDCreative, iframe });
-        };
+        });
 
         // handle errors
-        script.onerror = reject;
+        script.addEventListener('error', reject);
     });
 }
