+++ src/lib/runVPAID.js
@@ -13,7 +13,7 @@ export default async function runVPAID(creative, VPAIDSource, vastClient, ad) {
     const { VPAIDCreative, iframe } = await loadVPAID(VPAIDSource.fileURL, this.playerContainer);
     const VPAIDCreativeVersion = VPAIDCreative.handshakeVersion(SUPPORTED_VPAID_VERSION);
     const canSupportVPAID = isVPAIDVersionSupported(VPAIDCreativeVersion);
-    const { clientWidth: width, clientHeight: height } = this.playerContainer;
+    const { clientWidth: width, clientHeight: height } = this.container;
     const originalDimensions = { width, height };
     this.VASTTracker = vastTracker;
     if (canSupportVPAID) {
@@ -27,7 +27,7 @@ export default async function runVPAID(creative, VPAIDSource, vastClient, ad) {
         });
         this.__mountVideoElement();
         this.__renderControls();
-        this.__initVPAIDAd({ adParameters });
+        initVPAIDAd.call(this, { adParameters });
     }
 }
 
@@ -39,6 +39,10 @@ function attachVPAIDListeners() {
         this.__attachStoredListeners();
         this.__configureVPAID();
         this.__setReadyNextTick();
+        // Force the VPAID to cover the whole container
+        // This might not be the best thing to do tho
+        this.player.style.width = '100%';
+        this.player.style.height = '100%';
     });
     this.once('AdSizeChange', dimensions => {
         this.dimensions = dimensions;
@@ -55,4 +59,29 @@ function attachVPAIDListeners() {
         this.VPAIDWrapper = null;
         this.VASTData = null;
     });
+}
+
+let desiredBitrate;
+
+//const getViewMode = width => (width < 360 ? 'thumbnail' : 'normal');
+const viewMode = 'normal';
+
+function initVPAIDAd({ adParameters }) {
+    const { clientWidth: width, clientHeight: height } = this.container;
+    const creativeData = { AdParameters: adParameters };
+
+    const environmentVars = {
+        slot: this.playerContainer,
+        videoSlot: this.player,
+        videoSlotCanAutoplay: true
+    };
+    const initAdParameters = [
+        width,
+        height,
+        viewMode,
+        desiredBitrate,
+        creativeData,
+        environmentVars
+    ];
+    this.VPAIDWrapper.initAd(...initAdParameters);
 }
